double back-pressed added

+++ app/src/main/java/org/techtown/sns_project/SignUpActivity.java
@@ -35,6 +35,8 @@ public class SignUpActivity extends AppCompatActivity {
     private static final String TAG = "SignUpActivity";
     private RadioGroup radioGroup;
     private static boolean isNormal, isEnter;
+    private long backKeyPressedTime = 0;
+    private Toast terminate_guide_msg;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -113,15 +115,23 @@ public class SignUpActivity extends AppCompatActivity {
         Toast.makeText(this, msg, Toast.LENGTH_SHORT).show();
     }
 
-    // 로그아웃을 누른 뒤 회원가입창에서 다시 뒤로가기를 누르면 메인 액티비티 창이 또
-    // 나타나는 현상을 없애기 위함.
-    // 회원가입 창에서 뒤로가기를 누르면 앱을 종료하는 코드
     @Override
     public void onBackPressed() {
-        super.onBackPressed();
-        moveTaskToBack(true);
-        android.os.Process.killProcess(android.os.Process.myPid());
-        System.exit(1);
+
+        if (System.currentTimeMillis() > backKeyPressedTime + 1000) {
+            backKeyPressedTime = System.currentTimeMillis();
+            terminate_guide_msg = Toast.makeText(this, "\'뒤로\' 버튼을 한번 더 누르시면 종료됩니다.", Toast.LENGTH_SHORT);
+            terminate_guide_msg.show();
+            return;
+        }
+
+        if (System.currentTimeMillis() <= backKeyPressedTime + 1000) {
+            terminate_guide_msg.cancel();
+            moveTaskToBack(true);
+            android.os.Process.killProcess(android.os.Process.myPid());
+            System.exit(1);
+        }
+
     }
 
     View.OnClickListener onClickListener = new View.OnClickListener() {
